Pool: add getters for config & connection numbers
Adds `config()`, `connections()` and `idle_connections()` methods to
`Pool`. These methods simply return the current config object and the
number of connections open. This is very useful for health checking
the pool as a whole.

If, for example, `connections()` consistently returns a number less
than `config().min_idle().unwrap()` it means you are in trouble.
If it falls to zero then none of your servers is up. A good use case
for this is monitoring your servers.

While the same info is included when you debug print the pool,
which is very helpful during development to track down a bug,
exposing this info in a public method like this makes it feasible
to act on it in production.