Add `Pool#get_immediately`
This method behaves the same as `get`, except that it will not attempt
to establish new connections, and will not wait for a connection to
become available. This is useful for users who may wish to attempt to
parallelize a task over multiple connections, but only if there are
enough idle connections already available.