Fix rare race condition
Since the lock is dropped when try_get returns, we could run into the
case where get calls into try_get, there are no connections, but before
get can relock the lock and wait on the convar, all connections are
returned to the pool. Instead hold the lock after the check of the pool
state is made.