Add the ability to remove a connection from the pool
This changes how we handle unwind safety by adding another useful
feature, public API to permanently remove a connection from the pool.
When a connection is removed, the pool will immediately treat it the
same as getting back a broken connection. It will decrement its active
connection count, and immediately establish a new one in the pool if
this brings us below min_idle.

This feature allows us to prevent connections from being returned on
panic by adding a new hook to `ManageConnection`, which takes `&mut
PooledConnection`. It's assumed that any impl of `ManageConnection`
which implements `UnwindSafe` or `RefUnwindSafe` will override this hook
to restore any potentially broken invariants, or call `remove_from_pool`
if it can't.

We've introduced a new hook for this because we need to guarantee that
it will be called from the same thread the connection has dropped from,
which we don't guarantee for `has_broken` (and using that for this
purpose would be confusing)

I've structured things the way they are, because I want to eventually
have us start assuming connections will never be returned if they're
checked out for too long, so we'll need to pass the `Conn` around for
book keeping